training service refined for hybrid recommendations

+++ api/app/services/recommend.py
@@ -9,7 +9,7 @@ def model_recommend(user_id, user_type=None, top_n=6):
     with open("./app/services/cos_sim_matrix.pkl", "rb") as f:
         cos_sim_df = pickle.load(f)
     df_interactions = pd.read_parquet("./app/services/interactions.parquet")
-    user_id = int(user_id)
+    user_id = user_id
 
     if user_id in cos_sim_df.index:
         return recommend_for_logged_user(user_id, cos_sim_df, df_interactions, top_n)
@@ -34,10 +34,8 @@ def recommend_for_new_user(df_interactions, top_n):
     df_users = pd.read_pickle("./app/services/usuarios_preprocessados.pkl")
     default_features = get_default_user_features(kmeans)
     default_cluster = kmeans.predict([default_features])[0]
-    default_cluster = int(default_cluster)
+    default_cluster = default_cluster
     similar_users = df_users[df_users["cluster"] == default_cluster]["userId"]
-    df_users["userId"] = df_users["userId"].astype(int)
-    df_interactions["userId"] = df_interactions["userId"].astype(int)
     popular_articles = df_interactions[df_interactions["userId"].isin(similar_users)]
     recommended_articles = popular_articles.groupby("page").size().reset_index(name="popularity")
     recommended_articles = recommended_articles.sort_values("popularity", ascending=False)
